The server now host the ui.html and detect its IP
The server is now in charge of providing the ui.html page (on its root
url).

In order for everything to work, it required to detect its own IP, and a
big rework of the websockets on the client side.

Also corrected some errors in ui.html, as well as formatting.