Now use Jackson for server-side json messages
The messages exchanged by the webserver with the UI are no longer
constructed by hand, but are now generated from/to java objects using
the Jackson lib. This leads to a cleaner and better documented
communication protocol.

The only current limitation is that forms are still handled as string
and constructed by hand, to allow for different form generation
backends.
TODO: find an elegant solution to the form problem