Corrected single-thread bug in Process Dispatcher
Since ProcessDispatcher was mono-threaded, weird things could happen is
the tasks were handled in the same thread they were submitted. To avoid
this, new tasks submission is handled in a separate thread.

This of course has revealed that some classes were not thread-safe.
Notably, it seems that creating new servlets with jetty with
multi-threading is not a good idea...